@@ -41,48 +41,49 @@
         "@vap/shiki": "0.10.5",
         "fflate": "^0.8.2",
         "gifenc": "github:mattdesl/gifenc#64842fca317b112a8590f8fef2bf3825da8f6fe3",
-        "monaco-editor": "^0.50.0",
-        "nanoid": "^5.0.7",
+        "monaco-editor": "^0.52.2",
+        "nanoid": "^5.0.9",
         "virtual-merge": "^1.0.1"
     },
     "devDependencies": {
-        "@stylistic/eslint-plugin": "^2.6.1",
-        "@types/chrome": "^0.0.269",
-        "@types/diff": "^5.2.1",
-        "@types/lodash": "^4.17.7",
-        "@types/node": "^22.0.3",
-        "@types/react": "^18.3.3",
-        "@types/react-dom": "^18.3.0",
+        "@stylistic/eslint-plugin": "^2.12.1",
+        "@types/chrome": "^0.0.287",
+        "@types/diff": "^6.0.0",
+        "@types/lodash": "^4.17.14",
+        "@types/node": "^22.10.5",
+        "@types/react": "^19.0.2",
+        "@types/react-dom": "^19.0.2",
         "@types/yazl": "^2.4.5",
-        "diff": "^5.2.0",
+        "diff": "^7.0.0",
         "discord-types": "^1.3.26",
         "esbuild": "^0.15.18",
-        "eslint": "^9.8.0",
+        "eslint": "^9.17.0",
         "eslint-import-resolver-alias": "^1.1.2",
         "eslint-plugin-path-alias": "2.1.0",
-        "eslint-plugin-simple-header": "^1.1.1",
+        "eslint-plugin-react": "^7.37.3",
+        "eslint-plugin-simple-header": "^1.2.1",
         "eslint-plugin-simple-import-sort": "^12.1.1",
-        "eslint-plugin-unused-imports": "^4.0.1",
-        "highlight.js": "10.7.3",
+        "eslint-plugin-unused-imports": "^4.1.4",
+        "highlight.js": "11.7.0",
         "html-minifier-terser": "^7.2.0",
-        "moment": "^2.30.1",
-        "puppeteer-core": "^22.15.0",
+        "moment": "^2.22.2",
+        "puppeteer-core": "^23.11.1",
         "standalone-electron-types": "^1.0.0",
-        "stylelint": "^16.8.1",
+        "stylelint": "^16.12.0",
         "stylelint-config-standard": "^36.0.1",
-        "ts-patch": "^3.2.1",
-        "ts-pattern": "^5.3.1",
-        "tsx": "^4.16.5",
-        "type-fest": "^4.23.0",
-        "typescript": "^5.5.4",
-        "typescript-eslint": "^8.0.0",
-        "typescript-transform-paths": "^3.4.7",
+        "ts-patch": "^3.3.0",
+        "ts-pattern": "^5.6.0",
+        "tsx": "^4.19.2",
+        "type-fest": "^4.31.0",
+        "typescript": "^5.7.2",
+        "typescript-eslint": "^8.19.0",
+        "typescript-transform-paths": "^3.5.3",
         "zip-local": "^0.3.5"
     },
     "packageManager": "pnpm@9.1.0",
     "pnpm": {
         "patchedDependencies": {
-            "eslint@9.8.0": "patches/eslint@9.8.0.patch",
+            "eslint@9.17.0": "patches/eslint@9.17.0.patch",
             "eslint-plugin-path-alias@2.1.0": "patches/eslint-plugin-path-alias@2.1.0.patch"
         },
         "peerDependencyRules": {
