@@ -25,6 +25,7 @@ import { wordsToTitle } from "@utils/text";
 import definePlugin, { OptionType, PluginOptionsItem, ReporterTestable } from "@utils/types";
 import { findByPropsLazy } from "@webpack";
 import { Button, ChannelStore, Forms, GuildMemberStore, SelectedChannelStore, SelectedGuildStore, useMemo, UserStore } from "@webpack/common";
+import { ReactElement } from "react";
 
 interface VoiceState {
     userId: string;
@@ -289,7 +290,7 @@ export default definePlugin({
                 description: "Undeafen Message (only self for now)",
                 default: "{{USER}} undeafened"
             }
-        };
+        } satisfies Record<string, PluginOptionsItem>;
     },
 
     settingsAboutComponent({ tempSettings: s }) {
@@ -303,7 +304,7 @@ export default definePlugin({
             [],
         );
 
-        let errorComponent: React.ReactElement | null = null;
+        let errorComponent: ReactElement<any> | null = null;
         if (!hasVoices) {
             let error = "No narrator voices found. ";
             error += navigator.platform?.toLowerCase().includes("linux")
